get index to work with MultiCollection
