+++ Cargo.toml
@@ -17,7 +17,7 @@ sourmash = { git = "https://github.com/sourmash-bio/sourmash.git", branch = "rem
 serde_json = "1.0.125"
 niffler = "2.4.0"
 log = "0.4.22"
-env_logger = { version = "0.11.5", optional = true }
+env_logger = { version = "0.11.5" }
 simple-error = "0.3.1"
 anyhow = "1.0.86"
 zip = { version = "2.0", default-features = false }
