+++ src/index.rs
@@ -1,9 +1,12 @@
+use log::debug;
 use sourmash::index::revindex::RevIndex;
 use sourmash::index::revindex::RevIndexOps;
 use sourmash::prelude::*;
 use std::path::Path;
 
 use crate::utils::{load_collection, ReportType};
+use crate::utils::multicollection::MultiCollection;
+use sourmash::collection::Collection;
 
 pub fn index<P: AsRef<Path>>(
     siglist: String,
@@ -22,9 +25,14 @@ pub fn index<P: AsRef<Path>>(
         allow_failed_sigpaths,
     )?;
 
+    debug!("loaded collection from '{}' with len {}", siglist, collection.len());
+
+    let sigs = collection.load_sigs()?; // @CTB load into memory :sob:
+    let coll = Collection::from_sigs(sigs)?;
+
     let mut index = RevIndex::create(
         output.as_ref(),
-        collection.select(selection)?.try_into()?,
+        coll.select(selection)?.try_into()?,
         colors,
     )?;
