+++ src/mastiff_manysearch.rs
@@ -1,5 +1,6 @@
 /// mastiff_manysearch: mastiff-indexed version of manysearch.
 use anyhow::Result;
+use log::debug;
 use camino::Utf8PathBuf as PathBuf;
 use rayon::prelude::*;
 use std::sync::atomic;
@@ -26,6 +27,7 @@ pub fn mastiff_manysearch(
         bail!("'{}' is not a valid RevIndex database", index);
     }
     // Open database once
+    debug!("Opened revindex: '{index}')");
     let db = RevIndex::open(index, true, None)?;
 
     println!("Loaded DB");
@@ -73,6 +75,8 @@ pub fn mastiff_manysearch(
                             db.matches_from_counter(counter, minimum_containment as usize);
 
                         // filter the matches for containment
+                        debug!("FOUND: {} matches for {:?}", matches.len(),
+                               query_sig);
                         for (path, overlap) in matches {
                             let containment = overlap as f64 / query_size as f64;
                             if containment >= minimum_containment {
